refactor: move stack deserialization to model package
