+++ cmd/utils/stack.go
@@ -16,10 +16,8 @@ package utils
 import (
 	"fmt"
 	"path/filepath"
-	"strings"
 
 	oktetoErrors "github.com/okteto/okteto/pkg/errors"
-	oktetoLog "github.com/okteto/okteto/pkg/log"
 	"github.com/okteto/okteto/pkg/model"
 )
 
@@ -84,107 +82,4 @@ func LoadStackContext(stackPaths []string) (*model.ContextResource, error) {
 		}
 	}
 	return ctxResource, nil
-}
-
-// LoadStack loads an okteto stack manifest checking "yml" and "yaml"
-func LoadStack(name string, stackPaths []string) (*model.Stack, error) {
-	var resultStack *model.Stack
-
-	if len(stackPaths) == 0 {
-		stack, err := inferStack(name)
-		if err != nil {
-			return nil, err
-		}
-		resultStack = resultStack.Merge(stack)
-
-	} else {
-		for _, stackPath := range stackPaths {
-			if model.FileExists(stackPath) {
-				stack, err := getStack(name, stackPath)
-				if err != nil {
-					return nil, err
-				}
-
-				resultStack = resultStack.Merge(stack)
-				continue
-			}
-			return nil, fmt.Errorf("'%s' does not exist", stackPath)
-		}
-	}
-
-	if err := resultStack.Validate(); err != nil {
-		return nil, err
-	}
-	return resultStack, nil
-}
-
-func isPathAComposeFile(path string) bool {
-	base := filepath.Base(path)
-	return strings.HasPrefix(base, "docker-compose") || strings.HasPrefix(base, "okteto-compose")
-}
-
-func isDeprecatedExtension(stackPath string) bool {
-	base := filepath.Base(stackPath)
-	for _, deprecatedManifest := range deprecatedManifests {
-		if deprecatedManifest == base {
-			return true
-		}
-	}
-	return false
-}
-
-func getOverrideFile(stackPath string) (*model.Stack, error) {
-	extension := filepath.Ext(stackPath)
-	fileName := strings.TrimSuffix(stackPath, extension)
-	overridePath := fmt.Sprintf("%s.override%s", fileName, extension)
-	var isCompose bool
-	if model.FileExists(stackPath) {
-		if isPathAComposeFile(stackPath) {
-			isCompose = true
-		}
-		stack, err := model.GetStack("", overridePath, isCompose)
-		if err != nil {
-			return nil, err
-		}
-		return stack, nil
-	}
-	return nil, fmt.Errorf("override file not found")
-}
-func inferStack(name string) (*model.Stack, error) {
-	for _, possibleStackManifest := range possibleStackManifests {
-		manifestPath := filepath.Join(possibleStackManifest...)
-		if model.FileExists(manifestPath) {
-			stack, err := getStack(name, manifestPath)
-			if err != nil {
-				return nil, err
-			}
-
-			return stack, nil
-		}
-	}
-	return nil, oktetoErrors.UserError{
-		E:    fmt.Errorf("could not detect any stack file to deploy."),
-		Hint: "Try setting the flag '--file' pointing to your stack file",
-	}
-}
-
-func getStack(name, manifestPath string) (*model.Stack, error) {
-	var isCompose bool
-	if isDeprecatedExtension(manifestPath) {
-		deprecatedFile := filepath.Base(manifestPath)
-		oktetoLog.Warning("The file %s will be deprecated as a default stack file name in a future version. Please consider renaming your stack file to 'okteto-stack.yml'", deprecatedFile)
-	}
-	if isPathAComposeFile(manifestPath) {
-		isCompose = true
-	}
-	stack, err := model.GetStack(name, manifestPath, isCompose)
-	if err != nil {
-		return nil, err
-	}
-	overrideStack, err := getOverrideFile(manifestPath)
-	if err == nil {
-		oktetoLog.Info("override file detected. Merging it")
-		stack = stack.Merge(overrideStack)
-	}
-	return stack, nil
 }
